SEM_PA
Major changes have been made comparison to the first upload. Upon execution, the directory in which the particle analysis files are located is requested. Here, the main folder should be selected (not ‘Stub##’). The folder is automatically scanned for a *.csv file containing the data. The final report is saved as html file inside the selected directory. When a single particle is ‘tapped’, the micrographs of the particle is automatically loaded (which requires the report to be stored in the specified directory). Only single stubs are supported!